+++ package.json
@@ -5,10 +5,10 @@
     "description": "A simple command line utility to create and broadcast transactions.",
     "scripts": {
         "start": "node dist/index.js",
-        "message-sign": "node dist/message-sign.js",
-        "message-verify": "node dist/message-verify.js",
-        "wallet": "node dist/wallet-data.js",
-        "validator": "node dist/validator-data.js",
+        "message-sign": "node dist/actions/message-sign.js",
+        "message-verify": "node dist/actions/message-verify.js",
+        "wallet": "node dist/actions/wallet-data.js",
+        "validator": "node dist/actions/validator-data.js",
         "build": "del dist && tsc",
         "build:watch": "del dist && tsc -w",
         "debug": "node --inspect-brk dist/index.js",
