+++ src/actions/message-sign.ts
@@ -1,7 +1,7 @@
 import crypto from "crypto";
-import { makeApplication } from "./boot.js";
-import { loadConfig } from "./loader.js";
-import { makeIdentityFactories } from "./builder.js";
+import { makeApplication } from "../boot.js";
+import { loadConfig } from "../loader.js";
+import { makeIdentityFactories } from "../builder.js";
 
 const main = async () => {
     const config = loadConfig();
