+++ src/actions/validator-data.ts
@@ -1,7 +1,7 @@
 import { generateMnemonic } from "bip39";
-import { makeApplication } from "./boot.js";
-import { loadConfig } from "./loader.js";
-import { makeIdentityFactories } from "./builder.js";
+import { makeApplication } from "../boot.js";
+import { loadConfig } from "../loader.js";
+import { makeIdentityFactories } from "../builder.js";
 
 const main = async () => {
     const mnemonic = process.argv.length === 3 ? process.argv[2] : generateMnemonic(256);
