+++ src/actions/wallet-data.ts
@@ -1,6 +1,6 @@
-import { makeApplication } from "./boot.js";
-import { loadConfig } from "./loader.js";
-import { makeIdentityFactories } from "./builder.js";
+import { makeApplication } from "../boot.js";
+import { loadConfig } from "../loader.js";
+import { makeIdentityFactories } from "../builder.js";
 import { generateMnemonic } from "bip39";
 
 const main = async () => {
